+++ changelog.md
@@ -1,2 +1,4 @@
 # 1.0.0
-- Edit this file to change your mod's changelog.
+### Initial release
+
+- ![Add Icon](frame:GJ_completesIcon_001.png&scale:0.5) | **Added** base remote environment to fetch user badges